requirements.txt
fastapi==0.103.0
uvicorn==0.23.2
gunicorn==21.2.0
celery==5.3.1
redis==4.6.0
psycopg2-binary==2.9.6
transformers==4.42.4
torch==2.3.1
langgraph>=0.1.1
aiohttp==3.8.5
retry==0.9.2
spacy==3.7.2
